Moved ApplicationForm from containers->components
